Overall Project structure
Initial project structure has submodules under the "vendor" folder. "gtest" has been added here, since it'll most definitely be an integral part of development.
The "src" folder will contain the code for the library itself, while "sandbox" holds code for playing around with the usage of the library.
The "tests" folder should be fairly self-explanatory. It will contain tests for the project.
Other than that, there are also a few minor changes to other files.