Expand BigInt functionality and tests
This commit significantly expands the functionality of the BigInt class and provides improved test coverage. Several constructors have been added (not implemented), including direct initialisation from string view, a range of integer digits, and primitive integer types. Comparison and arithmetic operators are also now defined (not implemented). In addition, the tests for BigInt have been expanded to check various cases of (in)equality, and for self-equality with different inputs.