Add addition functionality and the basis for other operators
Added basic functionality to add and subtract digits of another BigInt. This can be used in conjunction with digitsCompare in suuri_core.hpp to implement, as an example, addition, which was done in this commit. The helper methods removeLeadingZeros was also added here, to accommodate some necessary functionality that will be used across multiple methods.

The constructor for BigInt from a builtin primitive was also fixed. The constructor wasn't technically working on specifically the suuri::base, which was brought up by the additions tests.